+++ src/config/settingsItems.ts
@@ -132,6 +132,10 @@ export const SETTINGS_CATEGORIES: SettingsCategory[] = [
       },
       { key: `${SETTINGS_MENU_KEY.proxies}.groupTestUrls`, label: 'groupTestUrls' },
       {
+        key: `${SETTINGS_MENU_KEY.proxies}.proxyFolderMode`,
+        label: 'proxyFolderMode',
+      },
+      {
         key: `${SETTINGS_MENU_KEY.proxies}.twoColumnProxyGroup`,
         label: 'twoColumnProxyGroup',
       },